inline a test config that is only used in one test and is very specifically named
